requirements.txt
biopython
tensorflow==2.0.0-alpha0